backend/pyproject.toml
[project]
name = "visionary-agent-protocol-backend"
version = "0.1.0"
description = "Backend services for Visionary Agent Protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.128.0",
  "uvicorn[standard]==0.38.0",
  "pydantic-settings==2.9.1",
  "python-multipart==0.0.19",
  "websockets==15.0.1",
  "openai==1.57.0",
  "numpy==2.2.6",
  "opencv-python-headless==4.10.0.84",
  "ultralytics==8.3.58"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "httpx==0.28.1",
  "ruff==0.8.4"
]
visionagents = [
  "vision-agents==0.3.5"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
